refactored emmbeder to work in the laser tokenizer package
