+++ pyproject.toml
@@ -13,6 +13,9 @@ requires-python = ">=3.8"
 dependencies = [
     'sacremoses==0.0.53',
     'sentencepiece==0.1.99',
+    'numpy==1.25.0',
+    'torch==2.0.1',
+    'fairseq==0.12.2',
 ]
 
 classifiers=[
@@ -64,4 +67,3 @@ testpaths = ["laser_encoders"]
 python_files = [
   "test_*.py",
 ]
-asyncio_mode = "auto"
